% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdm.R
\name{cvdm_ols_mr}
\alias{cvdm_ols_mr}
\title{Cross-Validated Difference in Means (CVDM) Test}
\usage{
cvdm_ols_mr(formula, data)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model to be fitted. The
details of model specification are given under 'Details.'}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}
}
\value{
An object to test whether a regression toward the mean, or
Ordinary Least Squares (OLS), or median regression (MR) is more appropriate. The object is the
Cross-Validated Johnson's t-test. A positive test statistic supports OLS
and a negative test statistic supports MR.
}
\description{
Applies cross-validated log-likelihood to test between
a regression toward the mean (ordinary least squares, or OLS) and a
median regression (MR). The output identifies the more appropriate model.
MR relies on the median instead of the mean, which might be more appropriate in
the case where outliers affect a regression to the mean. For more, see:
Harden, J. J., & Desmarais, B. A. (2011). Linear Models with Outliers:
Choosing between Conditional-Mean and Conditional-Median Methods.
State Politics & Policy Quarterly, 11(4), 371-389.
}
\details{
This function implements the cross-validated difference in means
(CVDM) test specifically for Ordinary Least Squares (OLS) and 'Median
Regression (MR). It returns the Cross Validated Johnson's Test (CVJT).
}
