% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdmtest.R
\name{cvdmtest}
\alias{cvdmtest}
\title{Cross-Validated Difference in Means (CVDM) Test}
\usage{
cvdmtest(formula, data, method1 = c("OLS", "MR", "RLM"),
  method2 = c("OLS", "MR", "RLM"), subset, na.action,
  singular.ok = TRUE)
}
\arguments{
\item{formula}{A formula object, with the dependent variable on the
left of a ~ operator, and the independent variables on the right.}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{method1}{A method to estimate the model. Currently takes
Ordinary Least Squares ("OLS"), Median Regression ("MR"), Robust Linear
Regression ("RLM") using M-estimation, and Robust Linear Regression using
MM-estimation ("RLM-MM"). The algorithm method used to compute the fit for the
median regression is the modified version of the Barrodale and Roberts algorithm
for l1-regression, which is the default by R package quantreg. See quantreg
qr function documentation for more deatils. Fitting for the robust regressions
is done by iterated re-weighted least squares (IWLS) and is taken from the
MASS package rlm function. The MM-estimation is the M-estimation with Tukey's
biweight initialized by a specific S-estimate. See MASS package rlm documentation
for details.}

\item{method2}{A method to estimate the model. Options
are same as for method1.}

\item{subset}{Expression indicating which subset of the rows of data should be
used in the fit. All observations are included by default.}

\item{na.action}{A missing-data filter function, applied to the model.frame,
after any subset argument has been used.}

\item{singular.ok}{Logical value indicating how to handle collinearity in the
model matrix. If \code{TRUE}, the program will automatically skip over columns
of the X matrix that are linear combinations of earlier columns. In this case
the coefficients for such columns will be NA, and the variance matrix will contain
zeros. For ancillary calculations, such as the linear predictor, the missing
coefficients are treated as zeros.}
}
\value{
An object of class \code{cvdm} computed by the cross-validated log likelihood
difference in means test (CVDM). The object is the Cross-Validated Johnson's t-test.
A positive test statistic supports the first method and a negative test statistic supports
the second. See \code{cvdm.object} for more details.
}
\description{
Applies cross-validated log-likelihood difference in means test to
compare two methods of estimating a formula. The output identifies the more
appropriate model.

In choosing between OLS and MR, please cite:

Harden, J. J., & Desmarais, B. A. (2011). Linear Models with Outliers:
Choosing between Conditional-Mean and Conditional-Median Methods.
State Politics & Policy Quarterly, 11(4), 371-389.

For other applications of the CVDM test, please cite:

Desmarais, B. A., & Harden, J. J. (2014). An Unbiased Model Comparison Test Using
Cross-Validation. Quality & Quantity, 48(4), 2155-2173.
}
\details{
This function implements the cross-validated difference in means (CVDM)
test between two methods of estimating a formula.
}
\examples{
\dontrun{
  set.seed(123456)
  b0 <- .2 # True value for the intercept
  b1 <- .5 # True value for the slope
  n <- 500 # Sample size
  X <- runif(n, -1, 1)

  Y <- b0 + b1 * X + rnorm(n, 0, 1) # N(0, 1 error)

  obj_cvdm <- cvdm(Y ~ X, data.frame(cbind(Y, X)), method1 = "OLS", method2 = "MR")
}
}
