% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVMF.R
\docType{data}
\name{cvmf}
\alias{cvmf}
\title{A function to compute the cross-validated median fit test
(CVMF).}
\format{An object of class \code{htest} of length 9.}
\usage{
cvmf
}
\value{
A function for the computation of cross-validated median fit test
(CVMF)
}
\description{
One of the main functions of the package.
Applies cross-validated log-likelihood to test between PLM and IRR.
}
\details{
Create function to compute the Cross-Validated Median Fit Test

\code{cvmf} returns the cross-validated median fit test

The functions contained in this file implement the cross-validated
median fit (CVMF) test. The function cvmf() tests between the partial
likelihood maximization (PLM) and the iteratively reweighted robust
(IRR) method of estimation for a given application of the Cox model.
}
\keyword{datasets}
