% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeLLtest.R
\docType{package}
\name{modeLLtest}
\alias{modeLLtest}
\alias{modeLLtest-package}
\title{modeLLtest}
\description{
modeLLtest
}
\section{modeLLtest Cross Validated Log Likelihood test functions}{

To use this package, decide which specification(s) of a model
and distributions you wish compare. The functions in this package
compare the fits of one model specification between a median
regression and ordinary least squares (cvdm()), between the fits
of one model specificiation between two estimationsof a Cox model
(cvmf()), and between two model specification and one distribution
(cvll()).
}

