% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.modeLLtest}
\alias{summary.modeLLtest}
\alias{summary.cvdm}
\alias{print.summary.cvdm}
\alias{summary.cvll}
\alias{print.summary.cvll}
\alias{summary.cvlldiff}
\alias{print.summary.cvlldiff}
\alias{summary.cvmf}
\alias{print.summary.cvmf}
\title{Summarizing modeLLtest Results}
\usage{
\method{summary}{cvdm}(object, ...)

\method{print}{summary.cvdm}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{cvll}(object, ...)

\method{print}{summary.cvll}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{cvlldiff}(object, ...)

\method{print}{summary.cvlldiff}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{cvmf}(object, ...)

\method{print}{summary.cvmf}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{cvdm}, \code{cvll},
\code{cvlldiff}, or \code{cvmf}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object returned by one of the \code{summary} methods.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
An object of class \code{summary.cvdm}, \code{summary.cvll},
  \code{summary.cvlldiff}, or \code{summary.cvmf}. Each is a list that
  contains the \code{call} (except for \code{cvlldiff}), the preferred
  method (\code{best}) where relevant, and a \code{table} of the test
  results. \code{summary.cvmf} also contains \code{plm_table} and
  \code{irr_table}, the coefficient tables for the partial likelihood
  and robust estimators.
}
\description{
\code{summary} methods for objects returned by \code{\link{cvdm}},
\code{\link{cvll}}, \code{\link{cvlldiff}}, and \code{\link{cvmf}}.
They collect the test results stored in the object into a list that can
be printed or used programmatically. No models are re-estimated.
}
\examples{
\donttest{
set.seed(123456)
X <- runif(200, -1, 1)
Y <- 0.2 + 0.5 * X + rnorm(200)
obj_cvdm <- cvdm(Y ~ X, data.frame(cbind(Y, X)),
                 method1 = "OLS", method2 = "MR")
summary(obj_cvdm)
}
}
\seealso{
\code{\link{cvdm_object}}, \code{\link{cvll_object}},
  \code{\link{cvlldiff_object}}, \code{\link{cvmf_object}}
}
