% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVDM.R
\name{CVDM}
\alias{CVDM}
\title{A function to compute the cross-validated difference in means test
 (CVDM).}
\usage{
CVDM(formula, data)
}
\arguments{
\item{CVDM}{The objects returned by computing the cross validated Johnson's
t-test (CVJT). #Check - why not Vuong here? does it compute this? Returns
CVJT and Vuong -- respective t-statistics, can be used as t or z stats.
Takes formula and data frame arguments. MORE ABOUT WHAT OBJECTS IT USES.
Both tests are fit(OLS)-fit(MR), such that negative values suport MR.}

\item{mu3hat}{The object returned by the estimation of skewness. It is used in
the estimation of the Johnson's t-test.}

\item{johnsons_t}{The object returned by the t-statistic adjustment for
skewness using the procedure suggested by Johnson (1978). It is estimated
using the mu3hat object and used to compute the CVDM test.}

\item{dlapl}{The object returned by the estimation of the centered Laplace
density. It is used in the estimation of the ll2 and cvll2 objects.}

\item{desingular}{The object returned by the estimation of MORE BOUT MATRIX.
It is used in the estimation of the cvll2 object.}

\item{ll2}{The object returned by the estimation of the individual likelihood
(or log likelihood). It is used in the construction of the CVDM test.}

\item{cvll2}{The object returned by the estimation of the cross-validated log
likelihood. It is estimated using the desingular object and is used to
construct the CVDM test.}
}
\value{
A function for the computation of Vuong and the Cross-Validated
 Johnson's t-test to test between OLS and MR
}
\description{
One of the main functions provided by the package.
}
\details{
Create function to compute the Cross-Validated Difference in Means Test

\code{CVDM} returns the cross-validated difference in means test

These are functions for the computation of Vuong and the Cross-Validated
Johnson's t-test. The main CVDM function tests between OLS and MR and returns
the CVJT and Vuong test -- respective t-statistics, can be used as t or z
stats. Both tests are fit(OLS)-fit(MR), such that negative values support MR.
For this to work properly, the arguments MORE.
}
