% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVDM.R
\name{cvdm}
\alias{cvdm}
\title{A function to compute the Cross-Validated Difference in Means Test.}
\usage{
cvdm(formula, data)
}
\arguments{
\item{cvdm}{The objects returned by computing the cross validated Johnson's
t-test (CVJT). MORE ABOUT WHAT OBJECTS IT USES. Negative values suport MR.}

\item{mu3hat}{The object returned by the estimation of skewness. It is
used in the estimation of the Johnson's t-test.}

\item{johnsons_t}{The object returned by the t-statistic adjustment for
skewness using the procedure suggested by Johnson (1978). It is estimated
using the mu3hat object and used to compute the CVDM test.}

\item{dlapl}{The object returned by the estimation of the centered Laplace
density. It is used in the estimation of the loglikes and cvloglikes objects.}

\item{desingular}{The object returned by the estimation of MORE BOUT MATRIX.
It is used in the estimation of the cvloglikes objects.}

\item{loglikes}{The object returned by the estimation of the individual
likelihood (or log likelihood). It is used in the construction of the CVDM test.}

\item{cvloglikes}{The object returned by the estimation of the
cross-validated log likelihood. It is estimated using the desingular object
and is used to construct the CVDM test.}
}
\value{
A function for the computation of Vuong and the Cross-Validated
Johnson's t-test to test between OLS and MR
}
\description{
One of the main functions of the package.
Applies cross-validated log-likelihood to test between OLS and MR.
}
\details{
Create function to compute the Cross-Validated Difference in Means (CVDM) Test

\code{CVDM} returns the cross-validated difference in means test

The functions contained in this file implement the cross-validated
difference in means (CVDM) test. The function cvdm() tests between
Ordinary Least Squares (OLS) and 'Median Regression (MR). It returns
the Cross Validated Johnson's Test (CVJT).For this to work properly,
the arguments MORE.
}
