% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdm.R
\name{cvdm}
\alias{cvdm}
\title{Cross-Validated Difference in Means (CVDM) Test}
\usage{
cvdm(formula, data)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model to be fitted. The
details of model specification are given under 'Details.'}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}
}
\value{
An object to test whether a regression toward the mean, or
Ordinary Least Squares (OLS), or median regression (MR) is more appropriate. The object is the
Cross-Validated Johnson's t-test. A positive test statistics support OLS
and a negative test statistics support MR.
}
\description{
Applies cross-validated log-likelihood to test if
a regression toward the mean (Ordinary Least Squares) or a
median regression (MR) is a more appropriate model.
}
\details{
This function implements the cross-validated difference in means
(CVDM) test. The function cvdm() tests between
Ordinary Least Squares (OLS) and 'Median Regression (MR). It returns
the Cross Validated Johnson's Test (CVJT). MORE about why and how.
For this to work properly,
the arguments MORE.
}
