% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVMF.R
\name{CVMF}
\alias{CVMF}
\title{A function to compute the cross-validated median fit test
(CVMF).}
\usage{
CVMF(formula, data, method = "breslow", trunc = 0.95)
}
\value{
A function for the computation of cross-validated median fit test
(CVMF)
}
\description{
One of the main functions provided by the package.
}
\details{
Create function to compute the Cross-Validated Median Fit Test

\code{CVMF} returns the cross-validated median fit test

The functions contained in this file implement the cross-validated
median fit (CVMF) test in Desmarais and Harden (YEAR).
The computation of this test statistic may take several minutes.
Given N observations in the data, the Cox model must be estimated N
times by both PLM and IRR (2N estimations).
}
