% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvll.R
\name{cvll}
\alias{cvll}
\title{Cross-Validated Log Likelihood (CVLL)}
\usage{
cvll(formula, data, method = c("OLS", "MR", "RLM"), subset, na.action,
  singular.ok = TRUE)
}
\arguments{
\item{formula}{A formula object, with the dependent variable on the
left of a ~ operator, and the independent variables on the right.}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{method}{The name of a method to estimate the model. Currently takes
Ordinary Least Squares ("OLS"), Median Regression ("MR"), Robust Linear
Regression ("RLM") using M-estimation, Robust Linear Regression using
MM-estimation ("RLM-MM"), partial likelihood maximization (PLM), and the
iteratively reweighted robust (IRR) method of estimation. The algorithm method
used to compute the fit for the median regression is the modified version of
the Barrodale and Roberts algorithm for l1-regression, which is the default
by R package quantreg. See quantreg qr function documentation for more deatils.
Fitting for the robust regressions is done by iterated re-weighted least squares
(IWLS) and is taken from the MASS package rlm function. The MM-estimation is
the M-estimation with Tukey's biweight initialized by a specific S-estimate.
See MASS package rlm documentation for details.}

\item{subset}{Expression indicating which subset of the rows of data should be
used in the fit. All observations are included by default.}

\item{na.action}{A missing-data filter function, applied to the model.frame,
after any subset argument has been used.}

\item{singular.ok}{Logical value indicating how to handle collinearity in the
model matrix. If \code{TRUE}, the program will automatically skip over columns
of the X matrix that are linear combinations of earlier columns. In this case
the coefficients for such columns will be NA, and the variance matrix will contain
zeros. For ancillary calculations, such as the linear predictor, the missing
coefficients are treated as zeros.}

\item{model}{The name of a statistical model to estimate.}
}
\value{
An object of class \code{cvll} computed by the cross-validated log likelihood
(CVLL). See \code{cvll.object} for more details.
}
\description{
Extracts the cross-validated log-likelihood from a method
of estimating a formula.
}
\details{
This function extracts a vector of the cross-validated log likelihood (CVLL)
from a method of estimating a formula.
}
