% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvll.R
\name{cvll}
\alias{cvll}
\title{Cross-Validated Log Likelihood (CVLL) Test}
\usage{
cvll(formula, data, method = c("OLS", "MR", "Poisson", "Bernoulli",
  "Logit", "Probit"), subset, na.action, singular.ok = TRUE, ...)
}
\arguments{
\item{formula}{A formula object, with the dependent variable on the
left of a ~ operator, and the independent variables on the right.}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{subset}{Expression indicating which subset of the rows of data should be
used in the fit. All observations are included by default.}

\item{na.action}{A missing-data filter function, applied to the model.frame,
after any subset argument has been used.}

\item{singular.ok}{Logical value indicating how to handle collinearity in the
model matrix. If \code{TRUE}, the program will automatically skip over columns
of the X matrix that are linear combinations of earlier columns. In this case
the coefficients for such columns will be NA, and the variance matrix will contain
zeros. For ancillary calculations, such as the linear predictor, the missing
coefficients are treated as zeros.}

\item{model}{The name of a statistical model to estimate.}
}
\value{
An object of class \code{cvll} computed by the cross-validated log likelihood
test (CVLL). See \code{cvll.object} for more details.
}
\description{
Applies cross-validated log-likelihood to test between
two methods of estimating a formula.
}
\details{
This function implements the cross-validated log likelihood (CVLL) test
between two methods of estimating a formula.
}
