% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvll.R
\name{cvll}
\alias{cvll}
\title{Cross-Validated Log Likelihood (CVLL)}
\usage{
cvll(formula, data, method = c("OLS", "MR", "RLM", "RLM-MM"), subset,
  na.action, ...)
}
\arguments{
\item{formula}{A formula object, with the dependent variable on the
left of a ~ operator, and the independent variables on the right.}

\item{data}{A data frame, list or environment (or object coercible by
as.data.frame to a data frame) containing the variables in the model.}

\item{method}{The name of a method to estimate the model. Currently takes
Ordinary Least Squares ("OLS"), Median Regression ("MR"), Robust Linear
Regression ("RLM") using M-estimation, Robust Linear Regression using
MM-estimation ("RLM-MM"), partial likelihood maximization (PLM), and the
iteratively reweighted robust (IRR) method of estimation. The algorithm method
used to compute the fit for the median regression is the modified version of
the Barrodale and Roberts algorithm for l1-regression, which is the default
by R package quantreg. See quantreg qr function documentation for more deatils.
Fitting for the robust regressions is done by iterated re-weighted least squares
(IWLS) and is taken from the MASS package rlm function. The MM-estimation is
the M-estimation with Tukey's biweight initialized by a specific S-estimate.
See MASS package rlm documentation for details.}

\item{subset}{Expression indicating which subset of the rows of data should be
used in the fit. All observations are included by default.}

\item{na.action}{A missing-data filter function, applied to the model.frame,
after any subset argument has been used.}

\item{model}{The name of a statistical model to estimate.}
}
\value{
An object of class \code{cvll} computed by the cross-validated log likelihood
(CVLL). See \code{cvdm_object} for more details.
}
\description{
Extracts the leave-one-out cross-validated log-likelihoods
from a method of estimating a formula.
}
\details{
This function extracts a vector of leave-one-out cross-validated
log likelihoods (CVLLs) from a method of estimating a formula.
Singular matrices during the leave-one-out cross-validation process
are skipped.
}
\examples{
\dontrun{
  set.seed(123456)
  b0 <- .2 # True value for the intercept
  b1 <- .5 # True value for the slope
  n <- 500 # Sample size
  X <- runif(n, -1, 1)

  Y <- b0 + b1 * X + rnorm(n, 0, 1) # N(0, 1 error)

  obj_cvll <- cvll(Y ~ X, data.frame(cbind(Y, X)), method = "OLS")
}
}
