% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlldiff.R
\name{cvlldiff}
\alias{cvlldiff}
\title{Cross-Validated Difference in Means (CVDM) Test with Vector Imputs}
\usage{
cvlldiff(vector1, vector2, df)
}
\arguments{
\item{vector1}{A numeric vector of cross-validated log-likelihoods.}

\item{vector2}{A numeric vector of cross-validated log-likelihoods.}

\item{df}{A value of the degrees of freedom in the models.}
}
\value{
An object of class \code{cvlldiff} computed by the cross-validated log likelihood
difference in means test (CVDM). The test statistic object is the Cross-Validated
Johnson's t-test. A positive test statistic supports the first method and a negative test
statistic supports the second.
}
\description{
Applies cross-validated log-likelihood to test between
two methods of estimating a formula. The output identifies the vector
from the more appropriate model.

Please cite:

Desmarais, B. A., & Harden, J. J. (2014). An Unbiased Model Comparison Test Using
Cross-Validation. Quality & Quantity, 48(4), 2155-2173.
}
\details{
This function implements the cross-validated difference in means (CVDM)
test between two vectors of cross-validated log-likelihoods.
}
