% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmf.R
\name{cvmf}
\alias{cvmf}
\title{A function to compute the cross-validated median fit test
(CVMF).}
\usage{
cvmf(formula, data, method = "efron", trunc = 0.95)
}
\arguments{
\item{cvmf}{The objects returned from computing the cross-validated median
fit test (CVMF). Uses the coxph() function from the survival package and
coxr() in the coxrobust package.}
}
\value{
A function for the computation of cross-validated median fit test
(CVMF).
}
\description{
One of the main functions of the package.
Applies cross-validated log-likelihood to test between PLM and IRR.
}
\details{
Create function to compute the Cross-Validated Median Fit Test

\code{cvmf} returns the cross-validated median fit test

The functions contained in this file implement the cross-validated
median fit (CVMF) test. The function cvmf() tests between the partial
likelihood maximization (PLM) and the iteratively reweighted robust
(IRR) method of estimation for a given application of the Cox model.
If outliers exist, the plm will be biased, but if no outliers exist,
the irr will have high variance.
}
