% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeLLtest.R
\docType{package}
\name{modeLLtest}
\alias{modeLLtest}
\title{modeLLtest Overview}
\description{
modeLLtest has three main functions to implement cross validated log likelihood
tests. To use this package, decide which specification(s) of a model
and distributions you wish compare. The function cvdm()
compares the fits of one model specification between a median
regression and ordinary least squares. The function cvmf() compares between
the fits of one model specificiation between two estimations of a Cox model.
The function cvll() extracts the leave-one-out cross-validated log-likelihoods
from a method of estimating a formula.
}
